Update coat-methods.R
Tried to implement 95%-CI in node_baplot() but did not succeed so far. I was able to determine the SE of a _weighted_ mean but was yet unable to determine the SE of a _weighted_ sample SD. The weighting makes it very complicated. Both quantities are needed to estimate the 95%-CI of the LoA. I will give it a last try tomorrow. If Siranush and I don't find a solution in the literature, i.e. an estimator for the SE of a _weighted_ sample SD, I will implement a bootstrapped solution. With a lot of time, the delta method could lead to a solution...